requirements.txt
python-telegram-bot==20.7
firebase-admin==6.2.0
python-dotenv==1.0.0
pytz==2023.3
pickle5==0.0.11
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
scikit-learn==1.3.0
Pillow==10.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dateutil==2.8.2
pycryptodome==3.19.0
emoji==2.8.0